Refactor formatter into full module tree with mod.rs
